Create useFeedsModel
